add check to ensure arguments exists in the inpute before updating said arguments get php unit tests to work after checking for existing values in the update product mutation

+++ server/src/Actions/UpdateProduct.php
@@ -50,49 +50,53 @@ class UpdateProduct
 
         /*Update images*/
         // Detach all product images (note this does not delete the image from the images table, only removed the association)
-        $product->images()->detach();
-        $images = json_decode($args["images"]);
+        if (isset($args["images"])) {
+            $images = json_decode($args["images"]);
+            $product->images()->detach();
 
-        // For each incoming image, if it already has an id, then update, otherwise create the new image and attach to product
-        $imagesNew = [];
-        foreach ($images as $image) {
 
-            // images from the front end with -1 as id will 'new' and needs to be created
-            if ($image->id >= 0) {
-                $imageFind = Image::find($image->id);
-                $imageFind->name = $image->name;
-                $imageFind->url = $image->url;
-                $imageFind->type = $image->type;
-                array_push($imagesNew, $imageFind);
-            } else {
-                $imageNew = new Image(["name" => $image->name, "url" => $image->url, "type" => $image->type]);
-                array_push($imagesNew, $imageNew);
+            // For each incoming image, if it already has an id, then update, otherwise create the new image and attach to product
+            $imagesNew = [];
+            foreach ($images as $image) {
+
+                // images from the front end with -1 as id will 'new' and needs to be created
+                if ($image->id >= 0) {
+                    $imageFind = Image::find($image->id);
+                    $imageFind->name = $image->name;
+                    $imageFind->url = $image->url;
+                    $imageFind->type = $image->type;
+                    array_push($imagesNew, $imageFind);
+                } else {
+                    $imageNew = new Image(["name" => $image->name, "url" => $image->url, "type" => $image->type]);
+                    array_push($imagesNew, $imageNew);
+                }
             }
+            $product->images()->saveMany($imagesNew); // make the new association of images to the product
         }
-        $product->images()->saveMany($imagesNew); // make the new association of images to the product
+
 
         /*Update tags*/
         // Detach all product tags (note this does not delete the tag from the tags table, only removed the association)
-        $product->tags()->detach();
-        $tags = json_decode($args["tags"]);
+        if (isset($args["tags"])) {
+            $tags = json_decode($args["tags"]);
+            $product->tags()->detach();
 
-        // For each incoming tag, if it already has an id, then update, otherwise create the new tag and attach to product
-        $tagsNew = [];
-        foreach ($tags as $tag) {
+            // For each incoming tag, if it already has an id, then update, otherwise create the new tag and attach to product
+            $tagsNew = [];
+            foreach ($tags as $tag) {
 
-            // tags from the front end with -1 as id will 'new' and needs to be created
-            if ($tag->id >= 0) {
-                $tagFind = Tag::find($tag->id);
-                $tagFind->name = $tag->name;
-                array_push($tagsNew, $tagFind);
-            } else {
-                $tagNew = new Tag(["name" => $tag->name]);
-                array_push($tagsNew, $tagNew);
+                // tags from the front end with -1 as id will 'new' and needs to be created
+                if ($tag->id >= 0) {
+                    $tagFind = Tag::find($tag->id);
+                    $tagFind->name = $tag->name;
+                    array_push($tagsNew, $tagFind);
+                } else {
+                    $tagNew = new Tag(["name" => $tag->name]);
+                    array_push($tagsNew, $tagNew);
+                }
             }
+            $product->tags()->saveMany($tagsNew); // make the new association of tags to the product
         }
-        $product->tags()->saveMany($tagsNew); // make the new association of tags to the product
-
-
         // Product details
         $productPrice = $args['price_cents'] ?? false;
         $productBaseCurrency = $args['price_currency'] ?? false;
